python-server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic[email]==2.10.3
simple-salesforce==1.12.6
requests==2.32.3
python-multipart==0.0.18
pyjwt==2.10.1
cryptography>=43.0.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
boto3==1.35.73
httpx==0.24.1